feat: refactor event more
