+++ src/constants/accessories.js
@@ -75,6 +75,8 @@ const accessoryUpgrades = [
   ["AGARIMOO_TALISMAN", "AGARIMOO_RING", "AGARIMOO_ARTIFACT"],
   ["BLOOD_DONOR_TALISMAN", "BLOOD_DONOR_RING", "BLOOD_DONOR_ARTIFACT"],
   ["LUSH_TALISMAN", "LUSH_RING", "LUSH_ARTIFACT"],
+  ["ANITA_TALISMAN", "ANITA_RING", "ANITA_ARTIFACT"],
+  ["PESTHUNTER_BADGE", "PESTHUNTER_RING", "PESTHUNTER_ARTIFACT"],
 ];
 
 const ignoredAccessories = [
@@ -103,6 +105,7 @@ const ignoredAccessories = [
   "WARDING_TRINKET",
   "RING_OF_BROKEN_LOVE",
   "GARLIC_FLAVORED_GUMMY_BEAR",
+  "GENERAL_MEDALLION",
 ];
 
 export const accessoryAliases = {
