+++ src/custom-resources.js
@@ -487,7 +487,16 @@ async function loadResourcePacks() {
           const apng = UPNG.encode(pngFrames, NORMALIZED_SIZE, NORMALIZED_SIZE, 0, pngDelays);
 
           await fs.writeFile(textureFile, Buffer.from(apng));
-          await execFile(apng2gif, [textureFile, textureFile.replace(".png", ".gif")]);
+
+          try {
+            if (fs.existsSync(textureFile.replace(".png", ".gif"))) {
+              await execFile(apng2gif, [textureFile, "-o", textureFile.replace(".png", ".gif")]);
+            } else {
+              await execFile(apng2gif, [textureFile, "-o", textureFile]);
+            }
+          } catch (error) {
+            console.log(error);
+          }
         }
       }
 
@@ -554,8 +563,8 @@ export async function getTexture(item, options) {
 
       if (
         options.ignore_id === false &&
-        (("skyblock_id" in texture && texture.skyblock_id != (item?.tag?.ExtraAttributes?.id ?? "")) ||
-          (!("skyblock_id" in texture) && item?.tag?.ExtraAttributes?.id !== undefined))
+        texture.skyblock_id === undefined &&
+        (!texture.match || texture.skyblock_id === item.tag?.ExtraAttributes?.id)
       ) {
         continue;
       }
