+++ src/stats.js
@@ -17,3 +17,5 @@ export * from "./stats/pets.js";
 export * from "./stats/skyblock-level.js";
 export * from "./stats/minions.js";
 export * from "./stats/rift.js";
+export * from "./stats/items.js";
+export * from "./stats/missing.js";
