+++ src/stats/items/equipment.js
@@ -1,0 +1,5 @@
+export function getEquipment(equipment) {
+  return {
+    equipment,
+  };
+}
