+++ views/sections/stats/items/armor.ejs
@@ -1,0 +1,60 @@
+<% function itemIcon(item, classes) { %>
+  <div <% if(item.texture_path){ %> style='background-image: url("<%= item.texture_path %>")' <% } %> class="
+    <%= classes.join(" ") %>
+      item-icon
+      <% if (isEnchanted(item)) { %> is-enchanted <% } %>
+      <% if (item.texture_path) { %> custom-icon <% } %>
+      <% if (item.Damage != 0) { %> icon-<%= item.id %>_0 <% } %>
+      icon-<%= item.id %>_<%= item.Damage %>
+    ">
+  </div>
+<% } %>
+
+<div class="stat-container stat-armor">
+  <a class="stat-anchor" id="Armor"></a>
+  <h2 class="stat-header">Armor</h2>
+  <div class="stat-content">
+    <% if (items.armor.armor.length == 0) { %>
+      <div class="no-access"><%= calculated.display_name %> doesn't have any armor equipped.</div>
+    <% } else { %>
+      <% if(items.armor.set_name) { %>
+        <p class="stat-raw-values">
+          <span class="stat-name">Set: </span><span class="stat-value piece-<%= items.armor.set_rarity %>-fg"><%= items.armor.set_name %></span>
+        </p>
+      <% } %>
+
+      <div class="pieces">
+        <% for(const item of items.armor.armor.slice().reverse()){ %>
+          <div tabindex="0" data-item-id="<%= item.itemId %>" class="rich-item piece piece-<%= item.rarity %>-bg <%= getRarityUpgradeClass(item) %>">
+            <% if (rarityOrder.indexOf(item.rarity) <= 4) { %>
+              <div class="piece-shine"></div>
+            <% } %>
+
+            <% itemIcon(item, ['piece-icon']); %>
+          </div>
+        <% } %>
+
+        <div data-bonus-stats="armor"></div>
+      </div>
+    <% } %>
+
+    <p class="stat-sub-header">Equipment</p>
+    <% if(Object.keys(items.equipment.equipment).length == 0){ %>
+      <div class="no-access"><%= calculated.display_name %> doesn't have any equipment equipped.</div>
+    <% } else { %>
+      <div class="pieces">
+        <% for(const item of items.equipment.equipment.slice().reverse()){ %>
+          <div tabindex="0" data-item-id="<%= item.itemId %>" class="rich-item piece piece-<%= item.rarity %>-bg <%= getRarityUpgradeClass(item) %>">
+            <% if(rarityOrder.indexOf(item.rarity) <= 4){ %>
+              <div class="piece-shine"></div>
+              <% } %>
+
+            <% itemIcon(item, ['piece-icon']); %>
+          </div>
+        <% } %>
+
+        <div data-bonus-stats="equipment"></div>
+      </div>
+    <% } %>
+  </div>
+</div>
