+++ views/sections/stats/items/wardrobe.ejs
@@ -1,0 +1,36 @@
+<% function itemIcon(item, classes) { %>
+  <div <% if(item.texture_path){ %> style='background-image: url("<%= item.texture_path %>")' <% } %> class="
+    <%= classes.join(" ") %>
+      item-icon
+      <% if (isEnchanted(item)) { %> is-enchanted <% } %>
+      <% if (item.texture_path) { %> custom-icon <% } %>
+      <% if (item.Damage != 0) { %> icon-<%= item.id %>_0 <% } %>
+      icon-<%= item.id %>_<%= item.Damage %>
+    ">
+  </div>
+<% } %>
+
+<% if (items.wardrobe.length > 0){ %>
+  <p class="stat-sub-header">Wardrobe</p>
+  <div class="pieces wardrobe">
+    <% for(const set of items.wardrobe){ %>
+      <div class="wardrobe-set">
+        <% for(const [index, item] of set.entries()) { %>
+          <% if (item) { %>
+            <div tabindex="0" data-item-id="<%= item.itemId %>" class="rich-item piece piece-<%= item.rarity %>-bg <%= getRarityUpgradeClass(item) %> index-<%= index %>">
+              <% if(rarityOrder.indexOf(item.rarity) <= 4){ %>
+                <div class="piece-shine"></div>
+              <% } %>
+              
+              <% itemIcon(item, ['piece-icon']); %>
+            </div>
+          <% } else { %>
+            <div class="piece armor-placeholder index-<%= index %>">
+              <div class="piece-icon item-icon custom-icon"></div>
+            </div>
+          <% } %>
+        <% } %>
+      </div>
+    <% } %>
+  </div>
+<% } %>
