+++ views/sections/stats/misc/end_island.ejs
@@ -9,66 +9,75 @@
 
   <p class="stat-raw-values">
     <!-- Most Damage -->
-    <%  let tooltip = "";
-    for (const key in calculated.misc.dragons.most_damage) {
-      if (key === "best") {
-        continue;
-      }
+    <% if (calculated.misc.dragons.most_damage !== undefined) { %>
+      <%  let tooltip = "";
+      for (const key in calculated.misc.dragons.most_damage) {
+        if (key === "best") {
+          continue;
+        }
  
-      tooltip += `<span class="stat-name">${helper.capitalizeFirstLetter(key)}: </span><span class="stat-value">${Math.floor(calculated.misc.dragons.most_damage[key]).toLocaleString()}</span><br>`;
-    } %>
+        tooltip += `<span class="stat-name">${helper.capitalizeFirstLetter(key)}: </span><span class="stat-value">${Math.floor(calculated.misc.dragons.most_damage[key]).toLocaleString()}</span><br>`;
+      } %>
 
-    <span data-tippy-content='<%= tooltip %>'>
-      <span class="stat-name">Most Damage: </span>
-      <span class="stat-value"><%= Math.floor(calculated.misc.dragons.most_damage.best).toLocaleString() %></span>
-    </span>
-    <br>
+      <span data-tippy-content='<%= tooltip %>'>
+        <span class="stat-name">Most Damage: </span>
+        <span class="stat-value"><%= Math.floor(calculated.misc.dragons.most_damage.best).toLocaleString() %></span>
+      </span>
+      <br>
+    <% } %>
 
     <!-- Fastest Kill -->
-    <%  tooltip = "";
-    for (const key in calculated.misc.dragons.fastest_kill) {
-      if (key === "best") {
-        continue;
-      }
+    <% if (calculated.misc.dragons.fastest_kill !== undefined) { %> 
+      <%  tooltip = "";
+      for (const key in calculated.misc.dragons.fastest_kill) {
+        if (key === "best") {
+          continue;
+        }
  
-      tooltip += `<span class="stat-name">${helper.capitalizeFirstLetter(key)}: </span><span class="stat-value">${moment.duration(calculated.misc.dragons.fastest_kill[key], "milliseconds").format("m:ss.SSS")}s</span><br>`;
-    } %>
+        tooltip += `<span class="stat-name">${helper.capitalizeFirstLetter(key)}: </span><span class="stat-value">${moment.duration(calculated.misc.dragons.fastest_kill[key], "milliseconds").format("m:ss.SSS")}s</span><br>`;
+      } %>
 
-    <span data-tippy-content='<%= tooltip %>'>
-      <span class="stat-name">Fastest Kill: </span>
-      <span class="stat-value"><%= moment.duration(calculated.misc.dragons.fastest_kill.best, "milliseconds").format("m:ss.SSS") %>s</span>
-    </span>
-    <br>
+      <span data-tippy-content='<%= tooltip %>'>
+        <span class="stat-name">Fastest Kill: </span>
+        <span class="stat-value"><%= moment.duration(calculated.misc.dragons.fastest_kill.best, "milliseconds").format("m:ss.SSS") %>s</span>
+      </span>
+      <br>
+    <% } %>
+    
     
     <!-- Last Hits (Kills) -->
-    <% tooltip = "";
-    for (const key in calculated.misc.dragons.kills) {
-      if (key === "total") {
-        continue;
-      }
+    <% if (calculated.misc.dragons.kills !== undefined) { %> 
+      <% tooltip = "";
+      for (const key in calculated.misc.dragons.kills) {
+        if (key === "total") {
+          continue;
+        }
      
-      tooltip += `<span class="stat-name">${helper.capitalizeFirstLetter(key)}: </span><span class="stat-value">${calculated.misc.dragons.kills[key].toLocaleString()}</span><br>`;
-    } %>
+        tooltip += `<span class="stat-name">${helper.capitalizeFirstLetter(key)}: </span><span class="stat-value">${calculated.misc.dragons.kills[key].toLocaleString()}</span><br>`;
+      } %>
 
-    <span data-tippy-content='<%= tooltip %>'>
-      <span class="stat-name">Last Hits: </span>
-      <span class="stat-value"><%= calculated.misc.dragons.kills.total.toLocaleString() %></span>
-    </span>
-    <br>
+      <span data-tippy-content='<%= tooltip %>'>
+        <span class="stat-name">Last Hits: </span>
+        <span class="stat-value"><%= calculated.misc.dragons.kills.total.toLocaleString() %></span>
+      </span>
+      <br>
+    <% } %>
     
     <!-- Deaths -->
-    <% tooltip = "";
-    for (const key in calculated.misc.dragons.deaths) {
-      if (key === "total") {
-        continue;
-      }
+    <% if (calculated.misc.dragons.deaths !== undefined) { %>
+      <% tooltip = "";
+      for (const key in calculated.misc.dragons.deaths) {
+        if (key === "total") {
+          continue;
+        }
      
-      tooltip += `<span class="stat-name">${helper.capitalizeFirstLetter(key)}: </span><span class="stat-value">${calculated.misc.dragons.deaths[key].toLocaleString()}</span><br>`;
-    } %>
-    <span data-tippy-content='<%= tooltip %>'>
-      <span class="stat-name">Deaths: </span>
-      <span class="stat-value"><%= calculated.misc.dragons.deaths.total.toLocaleString() %></span>
-    </span>
+        tooltip += `<span class="stat-name">${helper.capitalizeFirstLetter(key)}: </span><span class="stat-value">${calculated.misc.dragons.deaths[key].toLocaleString()}</span><br>`;
+      } %>
+      <span data-tippy-content='<%= tooltip %>'>
+        <span class="stat-name">Deaths: </span>
+        <span class="stat-value"><%= calculated.misc.dragons.deaths.total.toLocaleString() %></span>
+      </span>
+    <% } %>
   </p>
 <% } %>
