+++ views/sections/stats/skills.ejs
@@ -6,7 +6,7 @@
     <%- include('./skills/mining.ejs', {}) %>
 
     <!-- Farming -->
-    <%- include('./skills/farming.ejs', {}) %>
+    <%- include('./skills/farming.ejs', { getRarityUpgradeClass, rarityOrder }) %>
       
     <!-- Fishing -->
     <%- include('./skills/fishing.ejs', {}) %>
