+++ views/sections/stats/skills/mining.ejs
@@ -28,7 +28,18 @@ function formatNucleusCrystal(crystal, state) {
   return `<span style='color: var(--§${color});'>${formatted}</span>`
 }
 
-%>
+function itemIcon(item, classes) { %>
+  <div <% if(item.texture_path){ %> style='background-image: url("<%= item.texture_path %>")' <% } %> class="
+    <%= classes.join(" ") %>
+      item-icon
+      <% if (isEnchanted(item)) { %> is-enchanted <% } %>
+      <% if (item.texture_path) { %> custom-icon <% } %>
+      <% if (item.Damage != 0) { %> icon-<%= item.id %>_0 <% } %>
+      icon-<%= item.id %>_<%= item.Damage %>
+    ">
+  </div>
+<% } %>
+
 
 <div class="stat-mining">
   <div class="category-header">
